Supporting npm ci for installing packages with npm
This is an optional flag on the WithNpmPackageInstallation method (default to false) that when it's set to true it uses npm ci not npm install.

While it seems like a simple fix, it required a bunch of refactoring to cascade down for all package managers. Also found a problem with how it was determining the lockfile for different package managers, so that should also be fixed.

Unsure how to write tests as there's no API surface that we can get that exposes this info without writing an API surface explicitly for testing - which seems sub-optimal.

Fixes #95