Adding support for IEmbeddingGenerator off OllamaSharp
Overhaul of how the types are registered. Removed deprectation from the methods that register the IOllamaApiClient type as we'll use that as the base everywhere, and instead adding new methods that will register it as the IChatClient or IEmbeddingGenerator, depending on the desire.

Introduced tests to cover each of those code paths