First WIP at the resource command to redownload a model
Reused some of the stuff from the life cycle hook (made them static and public).

Should create a new base type like OllamaResourceCommandAnnotation that does a bunch of the setup of setup around when it can be used and connection to OllamaSharp so we can have other hooks like delete/push/etc.