Removing ollama tests
These tests aren't really needed, we should be using the tests in the AppHostTests file to cover the scenarios.

Testing volume persistance is outside of the scope that we can control - we can call the APIs, we can verify that the right annotations are placed, but if Aspire doesn't respect them or Docker doesn't persist the volume, it's not something we can 'fix'.