Using a new app host in the ollama tests and removing OpenWebUI
It looks like we keep disk space issues in GitHub Action runners (docs say they have 14gb of storage), so it's time to do some rethinking of Ollama in particular.

Ollama itself is a ~3gb image, then we were pulling two models, each 2gb or greater, then Open WebUI, which is ~4gb. Combine that with SQL, DAB, Meilisearch, and so on, and we're getting quite large with on-disk usage.

So with Ollama we're now using a custom app host for testing which slims down what's installed, only Ollama and tinyllama as a model (all up that should be around 4gb, not the more than 10gb we used to use)