Changing how Aspire.Hosting is referenced
It was being added to all projects under src, regardless of whether they are hosting or client integrations. Removed this from the Directory.Build.props and added it to the Hosting integrations only.

Had to update some dependencies across the client integrations as they were getting access to what they needed as transient via Aspire.Hosting