Dapr migration initial import (#378)
* Dropping files from Aspire repo

* Moving to the right namespace/folder naming

* Adding to solution and getting it to compile

* Adding in the Dapr tests from the Aspire repo

Had to remove the DaprSchemaTests file as we can't do schema tests (missing a lot of infrastructure from the Aspire repo).

Had to edit the DaprTests to not use EnvironmentVariableEvaluator, which we can't leverage as it uses some internal types from Aspire.Hosting. This means that our testing of the environment variables is slightly different, and the values we assert against are not the docker internal host endpoints, but the public endpoints