Making the OllamaSharp integration treated as preview
Since we depend on MEAI we trigger NU5104 on the OllamaSharp package when doing a release build as MEAI is a pre-release package. To combat this, we now can mark a package as preview for a release, which will mean that we can ship something that isn't considered stable or 'done' through MS Build properties.