Add Go Build Tags Support (#733)
* Add Go Build Tags Support

I updated GolangAppHostingExtension to support specifying Go build tags
to be used when building and running Go applications. Go build tags can
be used to include or exclude files from the build process based On
specific conditions. For example, when using Aspire for development, a
Go developer may use a build tag to enable special code for
OpenTelemetry instrumentation or enable debugging features.

Closes #730

* Revise Command-Line Building

I revised the way that I was building the command line for the `go run`
command based on review recommendations. The new approach should improve
readability of the code.