Specify the Path of the Golang Executable
I enhanced GolangAppHostingExtension to define a new override of
AddGolangApp that allows specifying the path of the Golang executable
relative to the working directory. This is useful for scenarios where
the Golang executable is in a `cmd` subdirectory, but the working
directory needs to be the root of the project for accessing resources
relative to the working directory at runtime. It is also useful when
there are more than one Golang executable in a Go module.

Resolves #754