Make SqlProject behave more like built in resources (#994)
* Make SqlProject behave more like built in resources

- Rework Sql Project resource to let Aspire handle the waiting, rather than handling it via a custom implementation

- Refactor Publishing to not `Exited` with an exit code rather than `FailedToStart` if the publishing process started, but then failed.

- Use Standard Resource states
  - Pending / Not Started --> Let aspire handle it
  - Publishing --> Running
  - Exceptions during the publish process now become `Exited`

- Add icon for database projects

- Add resource properties for SqlPackage projects - in particular the state which will show up on the dashboard.

In particular this involves removing any custom publishing of the `ResourceReadyEvent` as this explicitly called out as something that should not be done:

> Important: Developers should not manually publish the ResourceReadyEvent. Aspire manages the transition to the ready state based on the presence and outcome of health checks. Manually firing this event can interfere with the orchestration logic.
https://github.com/dotnet/aspire/blob/main/docs/specs/appmodel.md#resource-health

The key to getting aspire to do the waiting for us is to use `OnInitialize` to kick off our process.  See the talking clock example at https://github.com/dotnet/aspire/blob/main/docs/specs/appmodel.md#example-custom-resource---talking-clock for more details.

One thing I'm not entirely sure of is whether or not a `SqlProject` resource was intended to deploy multiple bacpacks from a signle resource.  The current implemention look.  If this is not intended, we can get rid of the `DacpackTargetAnnotation` and inline those properties on the resource itself.  If this was intended, we can make `DacpackTargetAnnotation` public.

* Try fix tests.

* Set resource start/finish timestamps.

* Add project path to dashboard.

* Remove unused file & restore ability to explicitly deploy database with explicit start.

* Tweak test.